Move usecases, boundaries and DTOs to a new usecases package
