+++ app/src/main/java/com/maximapps/maxim_weather/mainScreen/usecases/fetchforecastbycoordinates/FetchForecastByCoordinatesImpl.kt
@@ -1,10 +1,8 @@
-package com.maximapps.maxim_weather.mainScreen.domain.usecases
+package com.maximapps.maxim_weather.mainScreen.usecases.fetchforecastbycoordinates
 
 import androidx.annotation.RequiresPermission
 import com.maximapps.maxim_weather.common.SingleUseCase
-import com.maximapps.maxim_weather.mainScreen.domain.repositories.LocationRepository
-import com.maximapps.maxim_weather.mainScreen.domain.repositories.WeatherRepository
-import com.maximapps.maxim_weather.mainScreen.domain.models.WeatherData
+import com.maximapps.maxim_weather.mainScreen.usecases.models.WeatherData
 import io.reactivex.rxjava3.core.Single
 import javax.inject.Inject
 
@@ -12,11 +10,11 @@ typealias FetchForecastByCoordinates = SingleUseCase<Nothing, WeatherData>
 
 class FetchForecastByCoordinatesImpl @Inject constructor(
     private val locationRepository: LocationRepository,
-    private val weatherRepository: WeatherRepository
+    private val locationWeatherRepository: LocationWeatherRepository
 ) : FetchForecastByCoordinates {
 
     @RequiresPermission("android.permission.ACCESS_FINE_LOCATION")
     override operator fun invoke(params: Nothing?): Single<WeatherData> = locationRepository
         .getLocation()
-        .flatMap(weatherRepository::fetchForecastByCoordinates)
+        .flatMap(locationWeatherRepository::fetchForecastByCoordinates)
 }
