+++ app/src/main/java/com/maximapps/maxim_weather/mainScreen/usecases/fetchforecastbyname/FetchForecastByNameImpl.kt
@@ -1,20 +1,18 @@
-package com.maximapps.maxim_weather.mainScreen.domain.usecases
+package com.maximapps.maxim_weather.mainScreen.usecases.fetchforecastbyname
 
 import com.maximapps.maxim_weather.common.SingleUseCase
-import com.maximapps.maxim_weather.mainScreen.domain.models.WeatherData
-import com.maximapps.maxim_weather.mainScreen.domain.repositories.WeatherRepository
+import com.maximapps.maxim_weather.mainScreen.usecases.models.WeatherData
 import io.reactivex.rxjava3.core.Single
-import java.lang.IllegalArgumentException
 import javax.inject.Inject
 
 typealias FetchForecastByName = SingleUseCase<FetchForecastByNameImpl.Params, WeatherData>
 
 class FetchForecastByNameImpl @Inject constructor(
-    private val repository: WeatherRepository
+    private val cityWeatherRepository: CityWeatherRepository
 ) : FetchForecastByName {
     override operator fun invoke(params: Params?): Single<WeatherData> = when (params) {
         null -> Single.error(IllegalArgumentException("Parameter city name can't be null."))
-        else -> repository.fetchForecastByName(params.cityName)
+        else -> cityWeatherRepository.fetchForecastByName(params.cityName)
     }
 
     data class Params(
