+++ app/src/main/java/com/maximapps/maxim_weather/mainScreen/usecases/models/WeatherData.kt
@@ -1,4 +1,4 @@
-package com.maximapps.maxim_weather.mainScreen.domain.models
+package com.maximapps.maxim_weather.mainScreen.usecases.models
 
 data class WeatherData(
     val cityName: String = "",
